UI, auth, and warehouse improvements
- Add PageLoader component and refactor all loading screens to use it (20% smaller logo)
- Add registration enabled/disabled toggle to admin site settings with DB migration, page guard, action-level block, and disabled page with full i18n
- Fix location parent picker: add search combobox with full code path, fix top_down_unit as valid parent, fix narrow column headers in front elevation view
- Fix floating-point height validation bug in warehouse locations
- Fix public auth layout: center forms vertically, proper max-width on AuthCard, footer positioning, remove SmtpMessage from sign-up
- Update sign-up button text in header
