+++ apps/web/src/app/[locale]/(public)/loading.tsx
@@ -1,19 +1,5 @@
-"use client";
-
-import { useTranslations } from "next-intl";
-import { BrandLoader } from "@/components/branding";
+import { PageLoader } from "@/components/page-loader";
 
 export default function PublicLoading() {
-  const t = useTranslations("common");
-
-  return (
-    <div className="flex flex-1 items-center justify-center py-24">
-      <BrandLoader
-        variant="beacon_swap"
-        label={t("loading")}
-        showWordmark={false}
-        logoClassName="h-32 w-32"
-      />
-    </div>
-  );
+  return <PageLoader className="py-24" />;
 }
