+++ apps/web/src/app/qr/[token]/loading.tsx
@@ -1,14 +1,5 @@
-import { BrandLoader } from "@/components/branding";
+import { PageLoader } from "@/components/page-loader";
 
 export default function QRRedirectLoading() {
-  return (
-    <div className="flex min-h-screen items-center justify-center">
-      <BrandLoader
-        variant="beacon_swap"
-        label="Loading..."
-        showWordmark={false}
-        logoClassName="h-32 w-32"
-      />
-    </div>
-  );
+  return <PageLoader className="min-h-screen" />;
 }
